Move Settings related classes out of their impl package
Why would you want to export a impl package to the user anyway. `impl`
should be basically private and should not concern the user, since
"implementation specific". The original intent "implementations of
interfaces" is a little too verbose.
