+++ org.knime.knip.scijava.commands/src/org/knime/knip/scijava/commands/settings/SettingsModelTypePlugin.java
@@ -8,9 +8,9 @@ import org.scijava.plugin.SingletonPlugin;
  * A plugin which extends {@link SettingsModel} adapting capabilities.
  * 
  * <p>
- * SettingsModelTypePlugin plugins discoverable at runtime must implement this interface and be
- * annotated with @{@link Plugin} with attribute {@link Plugin#type()} =
- * {@link SettingsModelTypePlugin}.class.
+ * SettingsModelTypePlugin plugins discoverable at runtime must implement this
+ * interface and be annotated with @{@link Plugin} with attribute
+ * {@link Plugin#type()} = {@link SettingsModelTypePlugin}.class.
  * </p>
  * 
  * @author Jonathan Hale (University of Konstanz)
@@ -18,5 +18,7 @@ import org.scijava.plugin.SingletonPlugin;
  * @param <T>
  * @param <V>
  */
-public interface SettingsModelTypePlugin<T extends SettingsModel, V> extends SettingsModelType<T, V>, SingletonPlugin {
+public interface SettingsModelTypePlugin<T extends SettingsModel, V>
+		extends SettingsModelType<T, V>, SingletonPlugin {
+	// Marker interface
 }
