+++ org.knime.knip.scijava.commands/src/org/knime/knip/scijava/commands/settings/types/SettingsModelColorType.java
@@ -1,4 +1,4 @@
-package org.knime.knip.scijava.commands.settings.impl.types;
+package org.knime.knip.scijava.commands.settings.types;
 
 import java.awt.Color;
